requirements.txt
django
django-markupfield
Markdown
